Rename secrets.py to lambda_secrets_provider.py and update uses
