+++ shared/lambda_secrets_provider.py
